🐛 Interfaces that require tenant and user information use unified headers #355
