+++ frontend/services/agentConfigService.ts
@@ -1,17 +1,6 @@
 import { Tool, convertParamType } from '@/types/agentAndToolConst';
 import { API_ENDPOINTS } from './api';
-
-// 获取授权头的辅助函数
-const getAuthHeaders = () => {
-  const session = typeof window !== "undefined" ? localStorage.getItem("session") : null;
-  const sessionObj = session ? JSON.parse(session) : null;
-
-  return {
-    'Content-Type': 'application/json',
-    'User-Agent': 'AgentFrontEnd/1.0',
-    ...(sessionObj?.access_token && { "Authorization": `Bearer ${sessionObj.access_token}` }),
-  };
-};
+import { getAuthHeaders } from '@/lib/auth';
 
 /**
  * get tool list from backend
@@ -150,10 +139,7 @@ export const getCreatingSubAgentId = async (mainAgentId: string | null) => {
   try {
     const response = await fetch(API_ENDPOINTS.agent.getCreatingSubAgentId, {
       method: 'POST',
-      headers: {
-        ...getAuthHeaders(),
-        'Content-Type': 'application/json',
-      },
+      headers: getAuthHeaders(),
       body: JSON.stringify({ agent_id: mainAgentId }),
     });
 
@@ -203,9 +189,7 @@ export const updateToolConfig = async (
 
     const response = await fetch(API_ENDPOINTS.tool.update, {
       method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
+      headers: getAuthHeaders(),
       body: JSON.stringify({
         tool_id: toolId,
         agent_id: agentId,
@@ -244,9 +228,7 @@ export const searchToolConfig = async (toolId: number, agentId: number) => {
   try {
     const response = await fetch(API_ENDPOINTS.tool.search, {
       method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
+      headers: getAuthHeaders(),
       body: JSON.stringify({
         tool_id: toolId,
         agent_id: agentId
@@ -301,10 +283,7 @@ export const updateAgent = async (
   try {
     const response = await fetch(API_ENDPOINTS.agent.update, {
       method: 'POST',
-      headers: {
-        ...getAuthHeaders(),
-        'Content-Type': 'application/json',
-      },
+      headers: getAuthHeaders(),
       body: JSON.stringify({
         agent_id: agentId,
         name: name,
@@ -347,10 +326,7 @@ export const deleteAgent = async (agentId: number) => {
   try {
     const response = await fetch(API_ENDPOINTS.agent.delete, {
       method: 'DELETE',
-      headers: {
-        ...getAuthHeaders(),
-        'Content-Type': 'application/json',
-      },
+      headers: getAuthHeaders(),
       body: JSON.stringify({ agent_id: agentId }),
     });
 
@@ -380,10 +356,7 @@ export const exportAgent = async (agentId: number) => {
   try {
     const response = await fetch(API_ENDPOINTS.agent.export, {
       method: 'POST',
-      headers: {
-        ...getAuthHeaders(),
-        'Content-Type': 'application/json',
-      },
+      headers: getAuthHeaders(),
       body: JSON.stringify({ agent_id: agentId }),
     });
 
@@ -426,9 +399,7 @@ export const importAgent = async (agentId: string, agentInfo: any) => {
   try {
     const response = await fetch(API_ENDPOINTS.agent.import, {
       method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-      },
+      headers: getAuthHeaders(),
       body: JSON.stringify({ 
         agent_id: agentId, 
         agent_info: agentInfo 
@@ -464,10 +435,7 @@ export const searchAgentInfo = async (agentId: number) => {
   try {
     const response = await fetch(API_ENDPOINTS.agent.searchInfo, {
       method: 'POST',
-      headers: {
-        ...getAuthHeaders(),
-        'Content-Type': 'application/json',
-      },
+      headers: getAuthHeaders(),
       body: JSON.stringify({ agent_id: agentId }),
     });
