+++ frontend/services/configService.ts
@@ -1,18 +1,7 @@
 import { GlobalConfig } from '@/types/config';
 import { API_ENDPOINTS } from './api';
 import { ConfigStore } from '@/lib/config';
-
-// 获取授权头的辅助函数
-const getAuthHeaders = () => {
-  const session = typeof window !== "undefined" ? localStorage.getItem("session") : null;
-  const sessionObj = session ? JSON.parse(session) : null;
-
-  return {
-    'Content-Type': 'application/json',
-    'User-Agent': 'AgentFrontEnd/1.0',
-    ...(sessionObj?.access_token && { "Authorization": `Bearer ${sessionObj.access_token}` }),
-  };
-};
+import { getAuthHeaders } from '@/lib/auth';
 
 export class ConfigService {
   // Save global configuration to backend
@@ -20,10 +9,7 @@ export class ConfigService {
     try {
       const response = await fetch(API_ENDPOINTS.config.save, {
         method: 'POST',
-        headers: {
-          ...getAuthHeaders(),
-          'Content-Type': 'application/json',
-        },
+        headers: getAuthHeaders(),
         body: JSON.stringify(config),
       });
 
@@ -46,10 +32,7 @@ export class ConfigService {
     try {
       const response = await fetch(API_ENDPOINTS.config.load, {
         method: 'GET',
-        headers: {
-          ...getAuthHeaders(),
-          'Content-Type': 'application/json',
-        },
+        headers: getAuthHeaders(),
       });
       if (!response.ok) {
         const errorData = await response.json();
