+++ frontend/services/modelService.ts
@@ -2,6 +2,7 @@
 
 import { ModelOption, ModelType, ModelConnectStatus } from '../types/config'
 import { API_ENDPOINTS } from './api'
+import { getAuthHeaders } from '@/lib/auth'
 
 // API响应类型
 interface ApiResponse<T = any> {
@@ -29,24 +30,6 @@ export class ModelError extends Error {
   }
 }
 
-// Helper function to get authorization headers
-const getHeaders = () => {
-  return {
-    'Content-Type': 'application/json',
-  };
-};
-
-// 获取授权头的辅助函数
-const getAuthHeaders = () => {
-  const session = typeof window !== "undefined" ? localStorage.getItem("session") : null;
-  const sessionObj = session ? JSON.parse(session) : null;
-
-  return {
-    'Content-Type': 'application/json',
-    'User-Agent': 'AgentFrontEnd/1.0',
-    ...(sessionObj?.access_token && { "Authorization": `Bearer ${sessionObj.access_token}` }),
-  };
-};
 // Model service
 export const modelService = {
   // Get official model list
@@ -97,7 +80,7 @@ export const modelService = {
   getCustomModels: async (): Promise<ModelOption[]> => {
     try {
       const response = await fetch(API_ENDPOINTS.model.customModelList, {
-        headers: getHeaders()
+        headers: getAuthHeaders(),
       })
       const result: ApiResponse<any[]> = await response.json()
       
@@ -135,7 +118,7 @@ export const modelService = {
     try {
       const response = await fetch(API_ENDPOINTS.model.customModelCreate, {
         method: 'POST',
-        headers: getHeaders(),
+        headers: getAuthHeaders(),
         body: JSON.stringify({
           model_repo: "",
           model_name: model.name,
@@ -163,7 +146,7 @@ export const modelService = {
     try {
       const response = await fetch(API_ENDPOINTS.model.customModelDelete(displayName), {
         method: 'POST',
-        headers: getHeaders(),
+        headers: getAuthHeaders()
       })
       const result: ApiResponse = await response.json()
       if (result.code !== 200) {
@@ -181,7 +164,7 @@ export const modelService = {
       if (!displayName) return false
       const response = await fetch(API_ENDPOINTS.model.customModelHealthcheck(displayName), {
         method: "POST",
-        headers: getHeaders(),
+        headers: getAuthHeaders(),
         signal
       })
       const result: ApiResponse<{connectivity: boolean}> = await response.json()
@@ -211,7 +194,7 @@ export const modelService = {
     try {
       const response = await fetch(API_ENDPOINTS.model.verifyModelConfig, {
         method: "POST",
-        headers: getHeaders(),
+        headers: getAuthHeaders(),
         body: JSON.stringify({
           model_name: config.modelName,
           model_type: config.modelType,
