+++ frontend/services/promptService.ts
@@ -1,16 +1,5 @@
 import { API_ENDPOINTS } from './api';
-
-// 获取授权头的辅助函数
-const getAuthHeaders = () => {
-  const session = typeof window !== "undefined" ? localStorage.getItem("session") : null;
-  const sessionObj = session ? JSON.parse(session) : null;
-
-  return {
-    'Content-Type': 'application/json',
-    'User-Agent': 'AgentFrontEnd/1.0',
-    ...(sessionObj?.access_token && { "Authorization": `Bearer ${sessionObj.access_token}` }),
-  };
-};
+import { getAuthHeaders } from '@/lib/auth';
 
 /**
  * Prompt Generation Request Parameters
@@ -56,11 +45,7 @@ export const generatePromptStream = async (
   try {
     const response = await fetch(API_ENDPOINTS.prompt.generate, {
       method: 'POST',
-      headers: {
-        ...getAuthHeaders(),
-        'Content-Type': 'application/json',
-        'User-Agent': 'AgentFrontEnd/1.0',
-      },
+      headers: getAuthHeaders(),
       body: JSON.stringify(params),
     });
